A new class named sumArraySpecials have been added.
This function adds elements of an array. The elements corresponding to ending indices are added first. The elements are added in such a way that if the sum is more than or equal 10 the the unit digit is kept in the corresponding index of new array and the tenths place digit is moved or carried forward.